chore: 섹션 공통화 	modified:   src/screens/HomeScreen/CurrentForecastScreen/CurrentForecastScreen.tsx 	modified:   src/screens/HomeScreen/DailyForecastScreen/DailyForecastScreen.tsx 	modified:   src/screens/HomeScreen/DailyForecastScreen/_components/DailyForecastScreenHeader.tsx 	modified:   src/screens/HomeScreen/HourlyForecastScreen/HourlyForecastScreen.tsx 	modified:   src/screens/HomeScreen/HourlyForecastScreen/_components/HourlyForecastScreenHeader.tsx 	renamed:    src/screens/HomeScreen/CurrentForecastScreen/_components/forecastsGraphSection/DailyForecastsGraphSection.tsx -> src/screens/HomeScreen/_components/forecastsGraphSection/DailyForecastsGraphSection.tsx 	new file:   src/screens/HomeScreen/_components/forecastsGraphSection/DailyForecastsGraphSection.type.tsx 	renamed:    src/screens/HomeScreen/CurrentForecastScreen/_components/forecastsGraphSection/HourlyForecastsGraphSection.tsx -> src/screens/HomeScreen/_components/forecastsGraphSection/HourlyForecastsGraphSection.tsx 	new file:   src/screens/HomeScreen/_components/forecastsGraphSection/HourlyForecastsGraphSection.type.tsx 	renamed:    src/screens/HomeScreen/_components/forecastsGraph/ForecastsGraphDataPointComponent.tsx -> src/screens/HomeScreen/_components/forecastsGraphSection/customComponent/ForecastsGraphDataPointComponent.tsx 	renamed:    src/screens/HomeScreen/_components/forecastsGraph/ForecastsGraphLabelComponent.tsx -> src/screens/HomeScreen/_components/forecastsGraphSection/customComponent/ForecastsGraphLabelComponent.tsx 	renamed:    src/screens/HomeScreen/_components/forecastsGraph/ForecastsGraph.const.tsx -> src/screens/HomeScreen/_components/forecastsGraphSection/graph/ForecastsGraph.const.tsx 	renamed:    src/screens/HomeScreen/_components/forecastsGraph/ForecastsGraph.tsx -> src/screens/HomeScreen/_components/forecastsGraphSection/graph/ForecastsGraph.tsx 	renamed:    src/screens/HomeScreen/_components/forecastsGraph/ForecastsGraph.type.tsx -> src/screens/HomeScreen/_components/forecastsGraphSection/graph/ForecastsGraph.type.tsx 	renamed:    src/screens/HomeScreen/_components/forecastsGraph/ForecastsGraph.util.tsx -> src/screens/HomeScreen/_components/forecastsGraphSection/graph/ForecastsGraph.util.tsx 	new file:   src/screens/HomeScreen/_components/forecastsGraphSection/wrapper/ForecastsGraphSection.tsx 	new file:   src/screens/HomeScreen/_components/forecastsGraphSection/wrapper/ForecastsGraphSection.type.tsx 	renamed:    src/screens/HomeScreen/_components/forecastsGraph/ForecastsGraphSectionWrapper.tsx -> src/screens/HomeScreen/_components/forecastsGraphSection/wrapper/ForecastsGraphSectionWrapper.tsx 	renamed:    src/screens/HomeScreen/_components/forecastsGraph/ForecastsGraphSectionWrapper.type.tsx -> src/screens/HomeScreen/_components/forecastsGraphSection/wrapper/ForecastsGraphSectionWrapper.type.tsx
