+++ src/screens/HomeScreen/HourlyForecastScreen/HourlyForecastScreen.tsx
@@ -5,7 +5,7 @@ import { HOME_NAVIGATION_TEST_ID_LIST } from '@navigation/home/HomeNavigation.co
 import HomeScreenWrapper from '@screens/HomeScreen/_components/HomeScreenWrapper';
 import HourlyForecastScreenHeader from '@screens/HomeScreen/HourlyForecastScreen/_components/HourlyForecastScreenHeader';
 
-import type { ForecastsGraphSelectedIndex } from '@screens/HomeScreen/_components/forecastsGraph/ForecastsGraphSectionWrapper.type';
+import type { ForecastsGraphSelectedIndex } from '@screens/HomeScreen/_components/forecastsGraphSection/wrapper/ForecastsGraphSectionWrapper.type';
 
 export default function HourlyForecastScreen() {
   const [selectedIndex, setSelectedIndex] = useState<ForecastsGraphSelectedIndex>();
