+++ src/screens/HomeScreen/_components/forecastsGraphSection/HourlyForecastsGraphSection.type.tsx
@@ -1,0 +1,3 @@
+import type { ForecastsGraphSectionProps } from '@screens/HomeScreen/_components/forecastsGraphSection/wrapper/ForecastsGraphSection.type';
+
+export type HourlyForecastsGraphSectionProps = Omit<ForecastsGraphSectionProps, 'data' | 'children'>;
