+++ src/screens/HomeScreen/_components/forecastsGraphSection/graph/ForecastsGraph.const.tsx
