+++ src/screens/HomeScreen/_components/forecastsGraphSection/graph/ForecastsGraph.type.tsx
