+++ src/screens/HomeScreen/_components/forecastsGraphSection/wrapper/ForecastsGraphSectionWrapper.type.tsx
@@ -1,6 +1,6 @@
 import type { ViewProps } from 'react-native';
 
-import type { ForecastsGraphStyle } from '@screens/HomeScreen/_components/forecastsGraph/ForecastsGraph.type';
+import type { ForecastsGraphStyle } from '@screens/HomeScreen/_components/forecastsGraphSection/graph/ForecastsGraph.type';
 import type { CurrentForecastScreenSectionHeaderProps } from '@screens/HomeScreen/CurrentForecastScreen/_components/currentForecastScreenSectionHeader/CurrentForecastScreenSectionHeader.type';
 
 export type ForecastsGraphSelectedIndex = number | null;
@@ -9,7 +9,6 @@ export type ForecastsGraphSectionWrapperProps = ViewProps &
   Required<Pick<ForecastsGraphStyle, 'forecastsGraphSpacing'>> & {
     headerText?: CurrentForecastScreenSectionHeaderProps['text'];
     selectedIndex?: ForecastsGraphSelectedIndex;
-    onSelectedIndexChange?: (index: ForecastsGraphSelectedIndex) => void;
     hideHeader?: boolean;
     forecastsGraphContainerMargin: number;
   };
